Add checkbox unit test
